Fix truncating int issue when swiping slowly
Gestures.move_dist is now float so small errors in the swipe distance
(notably 3/4 finger swipes, since 2 finger swipes are fucked up by
rotates and scales) no longer accumulate.

+++ include/gestures.h
@@ -80,7 +80,7 @@ struct Gestures {
 	struct timeval tap_timeout;
 
 	int move_type;
-	int move_dist;
+	double move_dist;
 	int move_dir;
 	int move_drag;
 	int move_drag_dx;
